requirements.txt
Django==4.2
djangorestframework==3.14.0
psycopg2-binary==2.9.6
drf-yasg==1.21.5
django-rest-knox
rest_condition
python-decouple==3.7